# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Chat App",
        "title": "Chat App API",
        "contact": {
            "name": "Rama Jayapermana",
            "email": "jayapermanarama@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/api",
    "paths": {
        "/conversations/{conversationId}/messages": {
            "get": {
                "description": "Retrieve all messages in a conversation.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "Retrieve all messages in a conversation.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Conversation ID",
                        "name": "conversationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_raamaj_chat-app_internal_model.WebResponse-github_com_raamaj_chat-app_internal_model_MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_raamaj_chat-app_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raamaj_chat-app_internal_model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Send a message in a conversation.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "Send a message in a conversation.",
                "parameters": [
                    {
                        "description": "Message Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_raamaj_chat-app_internal_model.MessageRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Conversation ID",
                        "name": "conversationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_raamaj_chat-app_internal_model.WebResponse-github_com_raamaj_chat-app_internal_model_MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_raamaj_chat-app_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raamaj_chat-app_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users": {
            "post": {
                "description": "Create a new user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user.",
                "parameters": [
                    {
                        "description": "User Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_raamaj_chat-app_internal_model.RegisterUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_raamaj_chat-app_internal_model.WebResponse-github_com_raamaj_chat-app_internal_model_UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_raamaj_chat-app_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raamaj_chat-app_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "Retrieve user details by ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Retrieve user details by ID.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_raamaj_chat-app_internal_model.WebResponse-github_com_raamaj_chat-app_internal_model_UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_raamaj_chat-app_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raamaj_chat-app_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_raamaj_chat-app_internal_model.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {}
            }
        },
        "github_com_raamaj_chat-app_internal_model.MessageRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "sender_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_raamaj_chat-app_internal_model.MessageResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "conversation_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "sender_id": {
                    "type": "integer"
                },
                "sent_at": {
                    "type": "string"
                }
            }
        },
        "github_com_raamaj_chat-app_internal_model.PageMetadata": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total_item": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "github_com_raamaj_chat-app_internal_model.RegisterUserRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 100,
                    "example": "johndoe@example.com"
                },
                "password": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "123456"
                },
                "username": {
                    "type": "string",
                    "maxLength": 20,
                    "example": "johndoe"
                }
            }
        },
        "github_com_raamaj_chat-app_internal_model.UserResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2020-01-01T00:00:00+09:00"
                },
                "email": {
                    "type": "string",
                    "example": "johndoe@example.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "username": {
                    "type": "string",
                    "example": "johndoe"
                }
            }
        },
        "github_com_raamaj_chat-app_internal_model.WebResponse-github_com_raamaj_chat-app_internal_model_MessageResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_raamaj_chat-app_internal_model.MessageResponse"
                },
                "errors": {
                    "$ref": "#/definitions/github_com_raamaj_chat-app_internal_model.ErrorResponse"
                },
                "paging": {
                    "$ref": "#/definitions/github_com_raamaj_chat-app_internal_model.PageMetadata"
                }
            }
        },
        "github_com_raamaj_chat-app_internal_model.WebResponse-github_com_raamaj_chat-app_internal_model_UserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_raamaj_chat-app_internal_model.UserResponse"
                },
                "errors": {
                    "$ref": "#/definitions/github_com_raamaj_chat-app_internal_model.ErrorResponse"
                },
                "paging": {
                    "$ref": "#/definitions/github_com_raamaj_chat-app_internal_model.PageMetadata"
                }
            }
        }
    }
}